+++ crates/examples/Cargo.toml
@@ -110,6 +110,7 @@ chrono = "0.4"
 vbs = { workspace = true }
 sha2.workspace = true
 local-ip-address = "0.6"
+vec1 = { workspace = true }
 
 tracing = { workspace = true }
