+++ crates/testing/Cargo.toml
@@ -45,6 +45,7 @@ tracing = { workspace = true }
 vbs = { workspace = true }
 lru = { workspace = true }
 tagged-base64.workspace = true
+vec1 = { workspace = true }
 
 [target.'cfg(all(async_executor_impl = "tokio"))'.dependencies]
 tokio = { workspace = true }
