+++ crates/types/Cargo.toml
@@ -54,6 +54,7 @@ vbs = { workspace = true }
 displaydoc = { version = "0.2.3", default-features = false }
 dyn-clone = { git = "https://github.com/dtolnay/dyn-clone", tag = "1.0.17" }
 url = { workspace = true }
+vec1 = { workspace = true }
 
 [dev-dependencies]
 serde_json = { workspace = true }
