+++ crates/types/src/lib.rs
@@ -8,6 +8,7 @@ use light_client::StateVerKey;
 use tracing::error;
 use traits::signature_key::SignatureKey;
 use url::Url;
+use vec1::Vec1;
 
 use crate::utils::bincode_opts;
 pub mod consensus;
@@ -200,5 +201,5 @@ pub struct HotShotConfig<KEY: SignatureKey> {
     /// time to wait until we request data associated with a proposal
     pub data_request_delay: Duration,
     /// Builder API base URL
-    pub builder_url: Url,
+    pub builder_urls: Vec1<Url>,
 }
